+++ CHANGELOG.md
@@ -5,6 +5,12 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added `processList` graphql query to get the host's list of processes.
+
 ## [0.14.3] - 2023-09-04
 
 ### Changed
@@ -278,6 +284,7 @@ across our system.
 
 - An initial version.
 
+[Unreleased]: https://github.com/petabi/review-web/compare/0.14.3...main
 [0.14.3]: https://github.com/petabi/review-web/compare/0.14.2...0.14.3
 [0.14.2]: https://github.com/petabi/review-web/compare/0.14.1...0.14.2
 [0.14.1]: https://github.com/petabi/review-web/compare/0.14.0...0.14.1